some ios7 compatability stuff
git-svn-id: https://subversion.assembla.com/svn/kinvey/ios-library/trunk@8024 868178e2-d38b-4ba8-a603-825f91c87007

Former-commit-id: 9273559564c55bbcbe6b7019fb0c1cb582c3517d

+++ KinveyKit/KinveyKit/Source/KCSConnectionResponse.m
@@ -116,7 +116,7 @@
 #pragma mark - debug
 - (NSString *)debugDescription
 {
-    return [NSString stringWithFormat:@"%@ code: %d", [super debugDescription], _responseCode];
+    return [NSString stringWithFormat:@"%@ code: %ld", [super debugDescription], (long)_responseCode];
 }
 
 @end
